requirements.txt
fastapi
unicorn
requests
google-generativeai
python-dotenv